// src/charts/types.ts
export interface ChartConfig {
    width: number;
    height: number;
    leftPadding: number;
    rightPadding: number;
    topPadding: number;
    bottomPadding: number;
    chartWidth: number;
    chartHeight: number;
}

export interface ChartColors {
    onColor: string;
    offColor: string;
    bgColor: string;
    gridColor: string;
    textColor: string;
}